+++ src/game/tests.rs
@@ -485,6 +485,7 @@ mod vintage_cube_talon_gates_of_madara;
 mod vintage_cube_tamiyo;
 mod vintage_cube_tamiyo_inquisitive_student;
 mod vintage_cube_tear_asunder;
+mod vintage_cube_teferi_time_raveler;
 mod vintage_cube_tendrils;
 mod vintage_cube_territorial_kavu;
 mod vintage_cube_tersa_lightshatter;